security update and minor frontend changes

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/EventEndpoint.java
@@ -1,18 +1,22 @@
 package at.ac.tuwien.sepm.groupphase.backend.endpoint;
 
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.DetailedEventDto;
+import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.EmployerDto;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.EventInquiryDto;
-import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.SimpleEventDto;
+import at.ac.tuwien.sepm.groupphase.backend.endpoint.dto.ProfileDto;
 import at.ac.tuwien.sepm.groupphase.backend.endpoint.mapper.EventMapper;
+import at.ac.tuwien.sepm.groupphase.backend.entity.Employer;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Event;
+import at.ac.tuwien.sepm.groupphase.backend.entity.Profile;
 import at.ac.tuwien.sepm.groupphase.backend.service.EventService;
+import at.ac.tuwien.sepm.groupphase.backend.service.TokenService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.Authorization;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.access.AuthorizationServiceException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -30,11 +34,13 @@ public class EventEndpoint {
     private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     private final EventService eventService;
     private final EventMapper eventMapper;
+    private final TokenService tokenService;
 
     @Autowired
-    public EventEndpoint(EventMapper eventMapper, EventService eventService) {
+    public EventEndpoint(EventMapper eventMapper, EventService eventService, TokenService tokenService) {
         this.eventService = eventService;
         this.eventMapper = eventMapper;
+        this.tokenService = tokenService;
     }
 
     @PostMapping
@@ -44,7 +50,8 @@ public class EventEndpoint {
     @PreAuthorize("hasAuthority('ROLE_EMPLOYER')")
     public EventInquiryDto create(@Valid @RequestBody EventInquiryDto eventInquiryDto, @RequestHeader String authorization) {
         LOGGER.info("POST /api/v1/events/{}", eventInquiryDto);
-
+        Profile profile = tokenService.getProfileFromHeader(authorization);
+        eventInquiryDto.setEmployer(EmployerDto.EmployerDtoBuilder.aEmployerDto().withProfileDto(ProfileDto.ProfileDtoBuilder.aProfileDto().withId(profile.getId()).build()).build());
         return eventMapper.eventToEventInquiryDto(
             eventService.saveEvent(eventMapper.eventInquiryDtoToEvent(eventInquiryDto)));
 
@@ -72,8 +79,13 @@ public class EventEndpoint {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @ApiOperation(value = "Update an event", authorizations = {@Authorization(value = "apiKey")})
     @PreAuthorize("hasAuthority('ROLE_EMPLOYER')")
-    public void update(@Valid @RequestBody EventInquiryDto eventInquiryDto) {
+    public void update(@Valid @RequestBody EventInquiryDto eventInquiryDto, @RequestHeader String authorization) {
         LOGGER.info("PUT /api/v1/events/{}", eventInquiryDto);
+        Employer owner = eventService.findById(eventInquiryDto.getId()).getEmployer();
+        Employer employer = tokenService.getEmployerFromHeader(authorization);
+        if(owner != employer){
+            throw new AuthorizationServiceException(String.format("Keine Berchtigung um das Event \"%s\" zu editieren", eventInquiryDto.getTitle()));
+        }
         eventMapper.eventToEventInquiryDto(
             eventService.saveEvent(eventMapper.eventInquiryDtoToEvent(eventInquiryDto)));
 
@@ -83,8 +95,12 @@ public class EventEndpoint {
     @ResponseStatus(HttpStatus.OK)
     @ApiOperation(value = "Delete event and related data")
     @PreAuthorize("hasAuthority('ROLE_EMPLOYER')")
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable Long id, @RequestHeader String authorization) {
         LOGGER.info("DELETE /api/v1/events/{}", id);
+        Event e = eventService.findById(id);
+        if(!tokenService.getProfileFromHeader(authorization).getId().equals(e.getEmployer().getId())) {
+            throw new AuthorizationServiceException(String.format("Keine Berchtigung um das Event \"%s\" zu löschen", e.getTitle()));
+        }
         eventService.deleteEventById(id);
     }
